♻️Revamp templates code
one helper.ts
