+++ src/constants.ts
@@ -1,0 +1,4 @@
+export const CHAIN_ID = 3;
+export const TOKEN_TYPE_ETHER = 0;
+export const TOKEN_TYPE_ERC20 = 1;
+
