+++ src/constants_v1.ts
@@ -1,5 +1,4 @@
-export const CHAIN_ID = 3;
-export let MASK_CONTRACT_ADDRESS_LIST = new Array<string>(4);
+export let MASK_CONTRACT_ADDRESS_LIST = new Array<string>(5);
 
 // its placeholder to meet the declared array length
 MASK_CONTRACT_ADDRESS_LIST.push("0x0");
@@ -9,9 +8,9 @@ MASK_CONTRACT_ADDRESS_LIST.push("0x677f7BbA13108649ECFF068E8B3d55631327B83a");
 MASK_CONTRACT_ADDRESS_LIST.push("0xe0Cf2512237dFdE445eCDaFABffF98DB3AaCac76");
 // its placeholder to meet the declared array length
 MASK_CONTRACT_ADDRESS_LIST.push("0x0");
+// its placeholder to meet the declared array length
+MASK_CONTRACT_ADDRESS_LIST.push("0x0");
 
 export const GENESIS_TIMESTAMP = 1609372800; // UNIX timestamp
 export const GENESIS_TIMESTAMP_MASK = 1613088000; // UNIX timestamp
 
-export const TOKEN_TYPE_ETHER = 0;
-export const TOKEN_TYPE_ERC20 = 1;
