+++ src/constants_v2.ts
@@ -1,7 +1,3 @@
-export const CHAIN_ID = 3;
-
 export const GENESIS_TIMESTAMP = 1616976000; // UNIX timestamp
 export const GENESIS_TIMESTAMP_MASK = 1613088000; // UNIX timestamp
 
-export const TOKEN_TYPE_ETHER = 0;
-export const TOKEN_TYPE_ERC20 = 1;
