+++ src/mappingV2.ts
@@ -1,6 +1,7 @@
 import { Address, BigInt } from '@graphprotocol/graph-ts'
-import { fetchToken } from './helpers_v2'
-import { CHAIN_ID, GENESIS_TIMESTAMP } from './constants_v2'
+import { fetchToken } from './helpers'
+import { CHAIN_ID } from './constants'
+import { GENESIS_TIMESTAMP } from './constants_v2'
 import { FillSuccess, DestructSuccess, Fill_poolCall, SwapCall, SwapSuccess } from '../generated/ITO_V2/ITO_V2'
 import { PoolInfo, BuyInfo, DestructInfo, Pool, Seller, Buyer, Token, SellInfo } from '../generated/schema'
